+++ templates/withKibana/transactions.go
@@ -41,15 +41,15 @@ var Transactions = Object{
 			},
 			"gasLimit": Object{
 				"index": "false",
-				"type":  "long",
+				"type":  "double",
 			},
 			"gasPrice": Object{
 				"index": "false",
-				"type":  "long",
+				"type":  "double",
 			},
 			"gasUsed": Object{
 				"index": "false",
-				"type":  "long",
+				"type":  "double",
 			},
 			"hasOperations": Object{
 				"type": "boolean",
